requirements.txt
requests
discord.py
git+https://github.com/dolfies/discord.py-self@f062f9b8d39bb00a793d325e57d564046a461ce7#egg=selfcord.py